feat: Support multiple project files in a single release
Tries to update all detected project files in the targeted
repo/subfolder, moving all of them to the same version.

This might not actually be intended by the user, maybe this could be
made configurable?

Also the DefaultNix thing may only really work for my use case, so maybe
some sort of configurable custom thing would be the way forward?