Rework basically everything (#45)
The new implementation tries to be a lot less clever when figuring out what to update.

It now requires a configuration file (documentation TBD) that tells it what to do. It also drops the git2 crate and just runs git in a command.

BREAKING CHANGE: No more project detection, config is required
BREAKING CHANGE: pom.xml dropped for now

Co-authored-by: ci <server@bjoernerlwein.de>
Reviewed-on: http://git.ser.ver/bjoern/semantic-release/pulls/45
Co-authored-by: Björn Erlwein <bjoern@bjoernerlwein.de>
Co-committed-by: Björn Erlwein <bjoern@bjoernerlwein.de>